prototype: make the Scalmyth entity to go straight to the player if he can and also circle around the player.
The entity will try to go straight to the player, if there is no direct line of sight it will fallback to pathfinding.
The line of sight detection is better the vanilla,
but is only computed from the block center for every block in the entity width and height.
Also made the entity go around the player in a circle, it can get stuck pretty easily, currently.

The GoAround behavoiur can me costumized by editing the entity brain memory.
`enabled` if is false the go around behavoiur will not apply.
`points` how many segments the circle has.
`distance` how far away should be from the player.
`at` the current segment that is going to.

This was made in 12 Nov 2025