prototype: Mesh lighting, there are two modes.
Vertex Light when the `face_light` is false
Face Light when the `face_light` is true

KDebug visualization for the Vertex Normals and Face Normals in their respective mode.

Added a box.obj that is like a normal block.
The UV.y was fliped, to correspond with Blender UV.

Now the model is loaded one time, but the mesh is reconstructed every frame.
Also now the `scalmyth:mesh` will wait to be loaded before rendering itself.