Added collision to MashBlock and two experimental blocks to see if we can make multipart blocks.
wall_v1 is a block that can be rotated and is part of the house wall.
wall_v2 is a multipart block that is also for the house wall.
The problem is that the lighting is not calculated properly, this is why Minecraft doesn't have vertical slabs.

Also the Mesh collision the max collision size is one block that can have another block in every direction, so a double cross thing.